requirements.txt
anyio==4.9.0
certifi==2025.6.15
click==8.2.1
distro==1.9.0
fastapi==0.95.2
fire==0.7.0
fonttools==4.58.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
lxml==6.0.0
numpy==2.3.1
openai==1.93.0
opencv-python-headless==4.11.0.86
pdf2docx==0.5.8
pdf2image==1.17.0
pillow==11.3.0
pydantic==1.10.22
PyMuPDF==1.26.1
PyPDF2==3.0.1
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.6
redis==4.6.0
rq==1.15.1
sniffio==1.3.1
starlette==0.27.0
termcolor==3.1.0
tqdm==4.67.1
typing_extensions==4.14.0
uvicorn==0.23.1
